update Aventureiro class

+++ src/main/java/com/example/guildapi/repository/IAventureiroRepository.java
@@ -3,5 +3,5 @@ package com.example.guildapi.repository;
 import com.example.guildapi.model.Aventureiro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IAventureiroRepository extends JpaRepository<Aventureiro, Integer> {
+public interface IAventureiroRepository extends JpaRepository<Aventureiro, Long> {
 }
